refactor(cli):api
