+++ packages/amplication-cli/src/api/fields-list.ts
@@ -1,0 +1,46 @@
+import { ApolloClient, gql, NormalizedCacheObject } from '@apollo/client/core';
+
+import * as models from '../models';
+
+const GET_FIELDS = gql`
+  query getEntityFields(
+    $entityId: String!
+    $orderBy: EntityFieldOrderByInput
+    $whereName: StringFilter
+  ) {
+    entity(where: { id: $entityId }) {
+      id
+      appId
+      fields(where: { displayName: $whereName }, orderBy: $orderBy) {
+        id
+        displayName
+        name
+        dataType
+        required
+        searchable
+        description
+        properties
+      }
+    }
+  }
+`;
+
+export async function getFields(
+  client: ApolloClient<NormalizedCacheObject>,
+  entityId: string,
+  orderBy: string | undefined,
+  whereName: string | undefined
+): Promise<models.EntityField[]> {
+  const { data } = await client.query<{
+    entity: models.Entity;
+  }>({
+    query: GET_FIELDS,
+    variables: {
+      entityId,
+      orderBy,
+      whereName,
+    },
+  });
+
+  return data.entity.fields || [];
+}
