+++ packages/amplication-cli/src/api/index.ts
@@ -1,0 +1,10 @@
+export { getApp } from './apps-get';
+export { getApps } from './apps-list';
+export { createApp } from './apps-create';
+export { updateApp } from './apps-update';
+
+export { getEntities } from './entities-list';
+export { getEntity } from './entities-get';
+export { createEntity } from './entities-create';
+
+export { getFields } from './fields-list';
